+++ cmn-data-svc-server/src/main/java/org/acumos/cds/service/SolutionSearchServiceImpl.java
@@ -94,6 +94,7 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 	public Page<MLPSolution> findSolutions(Map<String, ? extends Object> queryParameters, boolean isOr,
 			Pageable pageable) {
 
+		Date beginDate = new Date();
 		Criteria criteria = sessionFactory.getCurrentSession().createCriteria(MLPSolution.class);
 		super.buildCriteria(criteria, queryParameters, isOr);
 
@@ -118,7 +119,7 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 
 		// Get a page of results and send it back with the total available
 		List<MLPSolution> items = criteria.list();
-		logger.debug(EELFLoggerDelegate.debugLogger, "findSolutions: result size={}", items.size());
+		logger.audit(beginDate, "findSolutions: result size={}", items.size());
 		return new PageImpl<>(items, pageable, count);
 	}
 
@@ -160,7 +161,6 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 			else
 				throw new AssertionFailure("Unexpected type: " + o.getClass().getName());
 		}
-
 		return new PageImpl<>(items, pageable, foms.size());
 	}
 
@@ -180,12 +180,9 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 			String[] ownerIds, String[] modelTypeCode, String[] accessTypeCode, String[] validationStatusCode,
 			String[] tags, Pageable pageable) {
 
+		Date beginDate = new Date();
 		// build the query using FOM to access child attributes
 		Criteria criteria = sessionFactory.getCurrentSession().createCriteria(MLPSolutionFOM.class);
-		criteria.createAlias("revisions", revAlias);
-		criteria.createAlias(revAlias + ".artifacts", artAlias);
-		criteria.createAlias("owner", ownerAlias);
-		criteria.createAlias("tags", tagAlias);
 		// Attributes on the solution
 		criteria.add(Restrictions.eq("active", active));
 		if (nameKeywords != null && nameKeywords.length > 0)
@@ -194,19 +191,25 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 			criteria.add(buildLikeListCriterion("description", descKeywords));
 		if (modelTypeCode != null && modelTypeCode.length > 0)
 			criteria.add(buildEqualsListCriterion("modelTypeCode", modelTypeCode));
-		// Attributes on other entities
-		if (accessTypeCode != null && accessTypeCode.length > 0)
-			criteria.add(buildEqualsListCriterion(revAlias + ".accessTypeCode", accessTypeCode));
-		if (validationStatusCode != null && validationStatusCode.length > 0)
-			criteria.add(buildEqualsListCriterion(revAlias + ".validationStatusCode", validationStatusCode));
-		if (ownerIds != null && ownerIds.length > 0)
+		if ((accessTypeCode != null && accessTypeCode.length > 0)
+				|| (validationStatusCode != null && validationStatusCode.length > 0)) {
+			criteria.createAlias("revisions", revAlias);
+			if (accessTypeCode != null && accessTypeCode.length > 0)
+				criteria.add(buildEqualsListCriterion(revAlias + ".accessTypeCode", accessTypeCode));
+			if (validationStatusCode != null && validationStatusCode.length > 0)
+				criteria.add(buildEqualsListCriterion(revAlias + ".validationStatusCode", validationStatusCode));
+		}
+		if (ownerIds != null && ownerIds.length > 0) {
+			criteria.createAlias("owner", ownerAlias);
 			criteria.add(Restrictions.in(ownerAlias + ".userId", ownerIds));
-		if (tags != null && tags.length > 0)
+		}
+		if (tags != null && tags.length > 0) {
+			// Tags are optional, so must use outer join
+			criteria.createAlias("tags", tagAlias, org.hibernate.sql.JoinType.LEFT_OUTER_JOIN);
 			criteria.add(Restrictions.in(tagAlias + ".tag", tags));
-
+		}
 		Page<MLPSolution> result = runSolutionFomQuery(criteria, pageable);
-		logger.debug(EELFLoggerDelegate.debugLogger, "findPortalSolutions: result size={}",
-				result.getNumberOfElements());
+		logger.audit(beginDate, "findPortalSolutions: result size={}", result.getNumberOfElements());
 		return result;
 	}
 
@@ -222,11 +225,14 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 	public Page<MLPSolution> findSolutionsByModifiedDate(boolean active, String[] accessTypeCode,
 			String[] validationStatusCode, Date date, Pageable pageable) {
 
+		Date beginDate = new Date();
 		// build the query using FOM to access child attributes
 		Criteria criteria = sessionFactory.getCurrentSession().createCriteria(MLPSolutionFOM.class);
+		// A solution should ALWAYS have revisions.
 		criteria.createAlias("revisions", revAlias);
+		// A revision should ALWAYS have artifacts
 		criteria.createAlias(revAlias + ".artifacts", artAlias);
-		criteria.createAlias(revAlias + ".solution", "sol");
+		// Attributes on the solution
 		criteria.add(Restrictions.eq("active", active));
 		if (accessTypeCode != null && accessTypeCode.length > 0)
 			criteria.add(Restrictions.in(revAlias + ".accessTypeCode", accessTypeCode));
@@ -244,8 +250,7 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 		criteria.add(itemModifiedAfter);
 
 		Page<MLPSolution> result = runSolutionFomQuery(criteria, pageable);
-		logger.debug(EELFLoggerDelegate.debugLogger, "findSolutionsByModifiedDate: result size={}",
-				result.getNumberOfElements());
+		logger.audit(beginDate, "findSolutionsByModifiedDate: result size={}", result.getNumberOfElements());
 		return result;
 	}